+++ plugins/data-management/package.json
@@ -39,6 +39,7 @@
   "dependencies": {
     "@gmod/ucsc-hub": "^0.1.6",
     "@mui/icons-material": "^5.0.1",
+    "@mui/x-data-grid": "^5.0.1",
     "clsx": "^1.1.0",
     "react-virtualized-auto-sizer": "^1.0.2",
     "react-vtree": "^3.0.0-beta.1",
